Update String model class to comply with JEP 254
Due to the implementation changes made to the standard Java String and
related classes in JEP 254 (Java 9 and later), JPF String model class
now fails to compile. This updates String model class to comply with the
internal API changes.

This make String model class to follow a internal representation similar
to the standard String class in Java 9 and later which uses byte array
plus an encoding-flag field instead of UTF-16 char array to represent a
String.

Fixes: #37