Suppress warnings in which Number constructor is used intendedly
Instantiating a new Long, Integer, Short or Byte object from a primitive
long, integer, short or byte argument is deprecated. It is recommended
that we use the more efficient static method valueOf() (introduced in
Java 5) instead, which caches objects for values between -128 and 127
inclusive.

However in JPF, there are few cases where the use of Number constructor
is intentional (So to bypass the caching). This commit suppress warnings
from appearing in such cases.

Issue: #47