Update JarClassFileContainer to comply with new class URI structure
This is related to the issue #50 which is about JPF failing to load
model classes from jpf-classes.jar due to different URI structure of the
.class files being introuduced in Java Module System.

The previous workaround introudced by the PR #51 is specifically for the
jpf-classes.jar And JPF will still fail to locate classes from the other
jar files.

Thus this reverts it. And make a implementation changes to
JarClassFileContainer which is JPF's universal continer for loading
classes from any jar file.