Refactor JPF_java_lang_String to fix invalid casting of value field
Most methods in JPF_java_lang_String fail as the `value` field have
changed from char[] to a byte[] since JEP 254.

Instead of retrieving the value field, and performing operations on that
value field to return a result (which is now complex as the value field
now being a byte[] and and having a coder which specifies the different
encoding), we turn JPF String object into a VM String object using
MJIEnv.getStringObject and then delegates the method call to that VM
object.