Delete JPF_java_lang_StringBuffer and JPF_java_lang_StringBuilder
Current implementations of JPF_java_lang_StringBuffer, and
JPF_java_lang_StringBuilder is highly obsolete as the value field now
being a byte[] instead of char[], and having a coder which specifies the
encoding. Also we are not able to use StringUTF16, StringLatin1 helper
classes to help us with encoding and decoding since they are being
package-private.

Neither StringBuffer nor StringBuilder have any native methods. So it
should be safe to remove these two.