Fix unnecessary assertion in String#getBytes(byte[], int, byte)
As it states "If two coders are different and the target is big enough,
invoker guarantees that the target is in UTF16"

So the invoker guarantees not to call when the source encoding is UTF-8
and the target is in LATIN-1. However the source and target can both be
in "UTF-8".

So the assertion that the coder is always UTF16 is wrong.