Modify Types.instanceOf to take a ClassInfo as LHS
This method's declaration currently reads `instanceOf(String, String)`,
taking both the LHS and RHS types as type signature strings.  As a
result, to use a `ClassInfo` object `ci` as the LHS, a caller must
convert it into a string via `ci.getType()` (see
`ElementInfo.instanceOf`), before the string eventually gets turned back
into a `ClassInfo` object (if it's a reference type) in the method body.

This patch changes the method declaration to read `instanceOf(ClassInfo,
String)`.  This change not only saves the back-and-forth, but also
addresses the case where the LHS class has been loaded by a non-default
class loader, in which case the original implementation fails to resolve
its type string back into a `ClassInfo`.

One downside is that if LHS is a nested array, the new implementation
will resolve a `ClassInfo` object for every "layer" of the array (via
`ClassInfo.getComponentClassInfo`), while the original implementation
"peels" the array via (cheaper) calls to `String.substring`.  I'm
guessing this performance degradation will not be noticeable in common
scenarios.

This patch also simplifies the implementation of `instanceOf`, ensures
it recognizes arrays as subtypes of Cloneable and Serializable (JLS
4.10.3), changes the names of a few variables to clarify they're
expected to contain type signatures, and adds unit tests (to
`ClassTest.java`, where the existing `testInstanceOf` unit test
resides).