Added model classes for java.nio.file and java.nio.charset packages #429 (#506)
* Found Potential Fix for #429

Seems like there needs to be model classes in the file and charset packages for java.nio. These are as small as I could make them

* Create NullCharsetTest.java

Developed UnitTest to address the error from #429

* Update NullCharsetTest.java

Should be "iae" and not "e"

* Fixed Failing Tests for NullCharSet

It turns out my implementation of Charset affected the behavior of underlying Gradle JPF tests. Removed this custom Charset implementation and instead kept four files to fix the root cause:
 - FileSystem.java
 - FileSystems.java
 - Pathy.java
 - spi/FileSystemProvider.java

All other files in the initial commit to PR were redundant.